+++ include/rudiments/private/client.h
@@ -3,15 +3,15 @@
 
 	protected:
 		char		**_connecterror();
-		long		_timeoutsec();
-		long		_timeoutusec();
-		unsigned long	_retrywait();
+		int32_t		_timeoutsec();
+		int32_t		_timeoutusec();
+		uint32_t	_retrywait();
 		uint32_t	_retrycount();
 
 		void	initialize(char *connecterror,
-					long timeoutsec,
-					long timeoutusec,
-					unsigned long retrywait,
+					int32_t timeoutsec,
+					int32_t timeoutusec,
+					uint32_t retrywait,
 					uint32_t retrycount);
 		void	clientClone(const client &c);
